Fix channels in bindat + maxone mea1k fixes

+++ spikeextractors/extractors/mea1krecordingextractor/mea1krecordingextractor.py
@@ -62,13 +62,13 @@ class Mea1kRecordingExtractor(RecordingExtractor):
             if np.any(np.diff(channel_idxs) < 0):
                 sorted_idx = np.argsort(channel_idxs)
                 recordings = self._signals[np.sort(channel_idxs), start_frame:end_frame]
-                return recordings[sorted_idx]
+                return recordings[sorted_idx].astype('float')
             else:
-                return self._signals[np.array(channel_idxs), start_frame:end_frame]
+                return self._signals[np.array(channel_idxs), start_frame:end_frame].astype('float')
         else:
             assert channel_ids in self.get_channel_ids()
             channel_idx = self.get_channel_ids().index(channel_ids)
-            return self._signals[np.array(channel_idx), start_frame:end_frame]
+            return self._signals[np.array(channel_idx), start_frame:end_frame].astype('float')
 
     @staticmethod
     def write_recording(recording, save_path):
